Refactor the meaty bits of the sync __init__.py into a utils module
