+++ enderchest/sync/utils.py
@@ -1,0 +1,60 @@
+"""Non-implementation-specific syncing utilities"""
+import getpass
+import os
+import socket
+from pathlib import Path
+from urllib.parse import ParseResult, unquote
+from urllib.request import url2pathname
+
+
+def get_default_netloc() -> str:
+    """Compile a netloc from environment variables, etc.
+
+    Returns
+    -------
+    str
+        The default netloc, which is {user}@{hostname}
+    """
+    return f"{getpass.getuser()}@{socket.gethostname()}".lower()
+
+
+def path_from_uri(uri: ParseResult) -> Path:
+    """Extract and unquote the path component of a URI to turn it into a pathlib.Path
+
+    h/t https://stackoverflow.com/a/61922504
+
+    Parameters
+    ----------
+    uri : ParseResult
+        The parsed URI to extract the path from
+
+    Returns
+    -------
+    Path
+        The path part of the URI as a Path
+    """
+    host = "{0}{0}{mnt}{0}".format(os.path.sep, mnt=uri.netloc)
+    return Path(os.path.abspath(os.path.join(host, url2pathname(unquote(uri.path)))))
+
+
+def render_remote(alias: str, uri: ParseResult) -> str:
+    """Render a remote to a descriptive string
+
+    Parameters
+    ----------
+    alias : str
+        The name of the remote
+    uri : ParseResult
+        The parsed URI for the remote
+
+    Returns
+    -------
+    str
+        {uri_string} [({alias})]}
+            (if different from the URI hostname)
+    """
+    uri_string = uri.geturl()
+
+    if uri.hostname != alias:
+        uri_string += f" ({alias})"
+    return uri_string
